This is an extremely large commit. It includes the creation of the following:
1) a dummy memory provider for easy prototyping
2) a mongo DB memory provider for storage
3) a few basic unit tests for the above
4) There was an exposed api for showing documents but this is probably non-functioning thanks to the refactoring the unit tests caused.
5) Library functions that also probably don't work due to the provider refactoring.

Why does 4 and 5 not work?

Because 4 and 5 was built around some express, mongo, node tutorial that was pretty cool, got data in and out of a database and seemed to work pretty well. It just used an extensive amount of callbacks. I felt somewhat livid about it because although callbacks are useful, it just lends itself to some nastiness. Plus, well, I started feeling sleezy about the lack of unit tests so I learned nodeunit.

Hah, and nodeunit really doesn't like the idea of passing callbacks around and it didn't seem as if it made for appropriate unit tests so I went back and refactored all of the providers to not use callbacks as I tested them. I'll go back and fix the exposed APIs and everything else required.