This change switches over to using mongojs vs mongodb native. For what?
Eh, I tried to do a bunch of simplification with these apis and finally discovered that the mongodb apis are async. So are the mongojs apis so that was a bit of a learning lesson, so I ended up putting a bunch of callbacks into the mongo provider code. Hah, thanks for breaking all the unit tests though. Will fix those in a later commit.

However, all of the service code works now, so I can forego all of the tests if I want and just continue building but that is a topic for later discussion. I also need to update the memory provider to work with this for much faster prototyping.